App_Prototype/Server/requirements.txt
Flask
Flask-PyMongo
pymongo
python-dotenv
flask-bcrypt
pyjwt